Created CommandFacade
