+++ src/php/PhelFacade.php
@@ -5,15 +5,13 @@ declare(strict_types=1);
 namespace Phel;
 
 use InvalidArgumentException;
-use Phel\Command\CommandFactoryInterface;
+use Phel\Command\CommandFacadeInterface;
 use Phel\Command\ReplCommand;
 use Phel\Command\RunCommand;
 use Phel\Command\TestCommand;
 use Phel\Compiler\GlobalEnvironment;
 use Phel\Exceptions\ExitException;
 use Phel\Runtime\RuntimeFactory;
-use Phel\Runtime\RuntimeInterface;
-use RuntimeException;
 
 final class PhelFacade
 {
@@ -35,18 +33,12 @@ Commands:
         Show this help message.
 
 HELP;
-    private const SUCCESS_CODE = 0;
-    private const FAILED_CODE = 1;
 
-    private string $currentDir;
-    private CommandFactoryInterface $commandFactory;
+    private CommandFacadeInterface $commandFacade;
 
-    public function __construct(
-        string $currentDir,
-        CommandFactoryInterface $commandFactory
-    ) {
-        $this->currentDir = $currentDir;
-        $this->commandFactory = $commandFactory;
+    public function __construct(CommandFacadeInterface $commandFacade)
+    {
+        $this->commandFacade = $commandFacade;
     }
 
     /**
@@ -74,9 +66,7 @@ HELP;
         $globalEnv = new GlobalEnvironment();
         RuntimeFactory::initialize($globalEnv)->loadNs('phel\core');
 
-        $this->commandFactory
-            ->createReplCommand($globalEnv)
-            ->run();
+        $this->commandFacade->executeReplCommand($globalEnv);
     }
 
     private function executeRunCommand(array $arguments): void
@@ -85,32 +75,11 @@ HELP;
             throw new InvalidArgumentException('Please provide a filename or namespace as argument!');
         }
 
-        $this->commandFactory
-            ->createRunCommand($this->loadVendorPhelRuntime())
-            ->run($arguments[0]);
+        $this->commandFacade->executeRunCommand($arguments[0]);
     }
 
     private function executeTestCommand(array $arguments): void
     {
-        $result = $this->commandFactory
-            ->createTestCommand($this->loadVendorPhelRuntime())
-            ->run($arguments);
-
-        ($result)
-            ? exit(self::SUCCESS_CODE)
-            : exit(self::FAILED_CODE);
-    }
-
-    private function loadVendorPhelRuntime(): RuntimeInterface
-    {
-        $runtimePath = $this->currentDir
-            . DIRECTORY_SEPARATOR . 'vendor'
-            . DIRECTORY_SEPARATOR . 'PhelRuntime.php';
-
-        if (!file_exists($runtimePath)) {
-            throw new RuntimeException('The Runtime could not be loaded from: ' . $runtimePath);
-        }
-
-        return require $runtimePath;
+        $this->commandFacade->executeTestCommand($arguments);
     }
 }
